refactor: extract Send sheet logic into feature module
