+++ src/features/send/screens/SendSheet.tsx
@@ -1,23 +1,20 @@
 import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { InteractionManager, Keyboard, Platform, View, type TextInput } from 'react-native';
 
-import { type StaticJsonRpcProvider } from '@ethersproject/providers';
-import { Wallet } from '@ethersproject/wallet';
 import { useRoute, type RouteProp } from '@react-navigation/native';
-import { useQueryClient } from '@tanstack/react-query';
 import { isEmpty, isEqual, isString } from 'lodash';
 import { useDebounce } from 'use-debounce';
-import { type Address } from 'viem';
 
 import { analytics } from '@/analytics';
+import { Column } from '@/components/layout';
 import { NoResults } from '@/components/list';
 import { NoResultsType } from '@/components/list/NoResults';
-import { useRainbowToastEnabled } from '@/components/rainbow-toast/useRainbowToastEnabled';
+import { SendAssetForm, SendAssetList, SendContactList, SendHeader } from '@/components/send';
+import { SheetActionButton } from '@/components/sheet';
 import { PROFILES } from '@/config/experimental';
 import useExperimentalFlag from '@/config/experimentalHooks';
 import { AssetType } from '@/entities/assetTypes';
 import { type ParsedAddressAsset } from '@/entities/tokens';
-import { TransactionStatus, type NewTransaction } from '@/entities/transactions';
 import { type UniqueAsset } from '@/entities/uniqueAssets';
 import { prefetchENSAvatar } from '@/features/ens/hooks/useENSAvatar';
 import { prefetchENSCover } from '@/features/ens/hooks/useENSCover';
@@ -27,21 +24,9 @@ import { debouncedFetchSuggestions } from '@/features/ens/utils/handlers';
 import { REGISTRATION_STEPS } from '@/features/ens/utils/helpers';
 import GasSpeedButton from '@/features/gas/components/GasSpeedButton';
 import useGas from '@/features/gas/hooks/useGas';
-import { parseGasParamsForTransaction } from '@/features/gas/utils/parseGas';
 import styled from '@/framework/ui/styled-thing';
-import { isNativeAsset } from '@/handlers/assets';
-import {
-  assetIsUniqueAsset,
-  buildTransaction,
-  createSignableTransaction,
-  estimateGasLimit,
-  getProvider,
-  isL2Chain,
-  resolveNameOrAddress,
-  type NewTransactionNonNullable,
-} from '@/handlers/web3';
-import { WrappedAlert as Alert } from '@/helpers/alert';
-import { convertAmountAndPriceToNativeDisplay, convertAmountFromNativeValue, formatInputDecimals, lessThan } from '@/helpers/utilities';
+import { assetIsUniqueAsset, buildTransaction, estimateGasLimit, resolveNameOrAddress } from '@/handlers/web3';
+import { convertAmountAndPriceToNativeDisplay, convertAmountFromNativeValue, formatInputDecimals } from '@/helpers/utilities';
 import { checkIsValidAddressOrDomain, checkIsValidAddressOrDomainFormat, isENSAddressFormat } from '@/helpers/validators';
 import useAccountSettings from '@/hooks/useAccountSettings';
 import useCoinListEditOptions from '@/hooks/useCoinListEditOptions';
@@ -53,24 +38,16 @@ import usePrevious from '@/hooks/usePrevious';
 import useSendableUniqueTokens from '@/hooks/useSendableUniqueTokens';
 import useSendSheetInputRefs from '@/hooks/useSendSheetInputRefs';
 import useUserAccounts from '@/hooks/useUserAccounts';
-import * as i18n from '@/languages';
 import { logger, RainbowError } from '@/logger';
-import { loadWallet, sendTransaction } from '@/model/wallet';
-import { setHardwareTXError } from '@/navigation/HardwareWalletTxNavigator';
 import { useNavigation } from '@/navigation/Navigation';
 import Routes from '@/navigation/routesNames';
 import { type RootStackParamList } from '@/navigation/types';
 import { type Contact } from '@/redux/contacts';
 import { rainbowTokenList } from '@/references/rainbow-token-list';
-import { interactionsCountQueryKey } from '@/resources/addys/interactions';
+import { getDefaultCheckboxes } from '@/screens/SendConfirmationSheet';
 import { useUserAssetsStore } from '@/state/assets/userAssets';
 import { useBackendNetworksStore } from '@/state/backendNetworks/backendNetworks';
 import { ChainId } from '@/state/backendNetworks/types';
-import { PAGE_SIZE } from '@/state/nfts/createNftsStore';
-import { useNftsStore } from '@/state/nfts/nfts';
-import { getNextNonce } from '@/state/nonces';
-import { addNewTransaction } from '@/state/pendingTransactions';
-import { executeFn, Screens, TimeToSignOperation } from '@/state/performance/performance';
 import { getWallets, useAccountAddress, useIsHardwareWallet } from '@/state/wallets/walletsStore';
 import { borders } from '@/styles';
 import { useTheme, type ThemeContextProps } from '@/theme/ThemeContext';
@@ -78,12 +55,10 @@ import deviceUtils from '@/utils/deviceUtils';
 import ethereumUtils from '@/utils/ethereumUtils';
 import isLowerCaseMatch from '@/utils/isLowerCaseMatch';
 import safeAreaInsetValues from '@/utils/safeAreaInsetValues';
-import { time } from '@/utils/time';
 
-import { Column } from '../components/layout';
-import { SendAssetForm, SendAssetList, SendContactList, SendHeader } from '../components/send';
-import { SheetActionButton } from '../components/sheet';
-import { getDefaultCheckboxes } from './SendConfirmationSheet';
+import { useSendChainState } from '../hooks/useSendChainState';
+import { useSendSubmit } from '../hooks/useSendSubmit';
+import { getSendSubmitButtonState } from '../utils/sendSheetUtils';
 
 const sheetHeight = deviceUtils.dimensions.height - (Platform.OS === 'android' ? 30 : 10);
 
@@ -122,7 +97,6 @@ const validateRecipient = (toAddress?: string, tokenAddress?: string) => {
 
 const validateRecipientDamagedState = (toAddress?: string) => {
   const wallets = getWallets();
-  // check for if the recipient is in a damaged wallet state and prevent
   if (wallets) {
     const internalWallet = Object.values(wallets).find(wallet =>
       wallet.addresses.some(address => isLowerCaseMatch(address.address, toAddress))
@@ -134,16 +108,8 @@ const validateRecipientDamagedState = (toAddress?: string) => {
   return true;
 };
 
-type OnSubmitProps = {
-  ens?: {
-    setAddress: boolean;
-    transferControl: boolean;
-    clearRecords: boolean;
-  };
-};
-
 export default function SendSheet() {
-  const { goBack, navigate } = useNavigation();
+  const { navigate } = useNavigation();
   const sortedAssets = useUserAssetsStore(state => state.legacyUserAssets);
   const isLoadingUserAssets = useUserAssetsStore(state => state.getStatus('isInitialLoad'));
   const {
@@ -167,7 +133,6 @@ export default function SendSheet() {
   const { nativeCurrency, chainId } = useAccountSettings();
   const accountAddress = useAccountAddress();
   const isHardwareWallet = useIsHardwareWallet();
-  const queryClient = useQueryClient();
 
   const { action: transferENS } = useENSRegistrationActionHandler({
     step: REGISTRATION_STEPS.TRANSFER,
@@ -181,8 +146,6 @@ export default function SendSheet() {
     isSufficientBalance: false,
     nativeAmount: '',
   });
-  const [currentChainId, setCurrentChainId] = useState<ChainId>(ChainId.mainnet);
-  const prevChainId = usePrevious(currentChainId);
   const [currentInput, setCurrentInput] = useState('');
 
   const { params } = useRoute<RouteProp<RootStackParamList, typeof Routes.SEND_SHEET>>();
@@ -195,13 +158,11 @@ export default function SendSheet() {
   const [nickname, setNickname] = useState('');
   const [selected, setSelected] = useState<ParsedAddressAsset | UniqueAsset | undefined>();
   const [maxEnabled, setMaxEnabled] = useState(false);
-  const { maxInputBalance, updateMaxInputBalance } = useMaxInputBalance();
 
   const [debouncedInput] = useDebounce(currentInput, 500);
   const [debouncedRecipient] = useDebounce(recipient, 500);
 
   const [isValidAddress, setIsValidAddress] = useState(!!recipientOverride);
-  const [currentProvider, setCurrentProvider] = useState<StaticJsonRpcProvider | undefined>(getProvider({ chainId: ChainId.mainnet }));
   const theme = useTheme();
   const { colors, isDarkMode } = theme;
 
@@ -213,6 +174,15 @@ export default function SendSheet() {
 
   const isUniqueAsset = assetIsUniqueAsset(selected);
 
+  const { currentChainId, currentProvider, isL2 } = useSendChainState({
+    accountChainId: chainId,
+    selectedAssetChainId: selected?.chainId,
+    startPollingGasFees,
+    stopPollingGasFees,
+  });
+
+  const { maxInputBalance, updateMaxInputBalance } = useMaxInputBalance();
+
   let colorForAsset = useColorForAsset(selected, undefined, false, true);
   const uniqueAssetColor = usePersistentDominantColorFromImage(isUniqueAsset ? selected?.images.lowResUrl : null) ?? colors.appleBlue;
   if (isUniqueAsset) {
@@ -220,16 +190,11 @@ export default function SendSheet() {
   }
 
   const isENS = selected?.type === AssetType.ens;
-
   const ensName = isENS ? selected?.name : '';
   const ensProfile = useENSProfile(ensName, {
     enabled: isENS,
     supportedRecordsOnly: false,
   });
-
-  const isL2 = useMemo(() => {
-    return isL2Chain({ chainId: currentChainId });
-  }, [currentChainId]);
 
   const sendUpdateAssetAmount = useCallback(
     (newAssetAmount: string) => {
@@ -304,41 +269,6 @@ export default function SendSheet() {
     sendUpdateSelected,
     updateMaxInputBalance,
   ]);
-
-  useEffect(() => {
-    // We can start fetching gas prices
-    // after we know the network that the asset
-    // belongs to
-    if (prevChainId !== currentChainId) {
-      InteractionManager.runAfterInteractions(() => {
-        startPollingGasFees(currentChainId);
-      });
-    }
-  }, [startPollingGasFees, prevChainId, currentChainId]);
-
-  // Stop polling when the sheet is unmounted
-  useEffect(() => {
-    return () => {
-      InteractionManager.runAfterInteractions(() => {
-        stopPollingGasFees();
-      });
-    };
-  }, [currentChainId, stopPollingGasFees]);
-
-  useEffect(() => {
-    const assetChainId = selected?.chainId;
-    if (assetChainId && (assetChainId !== currentChainId || !currentChainId || prevChainId !== currentChainId)) {
-      if (chainId === ChainId.goerli) {
-        setCurrentChainId(ChainId.goerli);
-        const provider = getProvider({ chainId: ChainId.goerli });
-        setCurrentProvider(provider);
-      } else {
-        setCurrentChainId(assetChainId);
-        const provider = getProvider({ chainId: currentChainId });
-        setCurrentProvider(provider);
-      }
-    }
-  }, [currentChainId, chainId, prevChainId, selected?.chainId]);
 
   const onChangeNativeAmount = useCallback(
     (newNativeAmount: string) => {
@@ -426,333 +356,38 @@ export default function SendSheet() {
     [accountAddress, amountDetails.assetAmount, currentChainId, currentProvider, selected, toAddress, updateTxFee]
   );
 
-  const onSubmit = useCallback(
-    async ({ ens }: OnSubmitProps = {}) => {
-      if (!selected || !currentProvider) return;
-
-      const wallet = await executeFn(loadWallet, {
-        operation: TimeToSignOperation.KeychainRead,
-        screen: isENS ? Screens.SEND_ENS : Screens.SEND,
-      })({
-        provider: currentProvider,
-        timeTracking: {
-          screen: isENS ? Screens.SEND_ENS : Screens.SEND,
-          operation: TimeToSignOperation.Authentication,
-        },
-      });
-      if (!wallet) return;
-
-      const currentChainIdNetwork = useBackendNetworksStore.getState().getChainsName()[currentChainId ?? ChainId.mainnet];
-
-      const validTransaction = isValidAddress && amountDetails.isSufficientBalance && isSufficientGas && isValidGas;
-      if (!selectedGasFee?.gasFee?.estimatedFee || !validTransaction) {
-        logger.error(new RainbowError(`[SendSheet]: preventing tx submit because selectedGasFee is missing or validTransaction is false`), {
-          selectedGasFee,
-          validTransaction,
-          isValidGas,
-        });
-        return false;
-      }
-
-      let submitSuccess = false;
-      let updatedGasLimit = null;
-
-      // Attempt to update gas limit before sending ERC20 / ERC721
-      if (!isUniqueAsset && selected && !isNativeAsset(selected.address, currentChainId)) {
-        try {
-          // Estimate the tx with gas limit padding before sending
-          updatedGasLimit = await estimateGasLimit(
-            {
-              address: accountAddress,
-              amount: Number(amountDetails.assetAmount),
-              asset: selected,
-              recipient: toAddress,
-            },
-            true,
-            currentProvider,
-            currentChainId
-          );
-
-          if (updatedGasLimit && !lessThan(updatedGasLimit, gasLimit)) {
-            if (useBackendNetworksStore.getState().getNeedsL1SecurityFeeChains().includes(currentChainId)) {
-              updateTxFeeForOptimism(updatedGasLimit);
-            } else {
-              updateTxFee(updatedGasLimit, null);
-            }
-          }
-          // eslint-disable-next-line no-empty
-        } catch (e) {}
-      }
-
-      let nextNonce;
-
-      if (isENS && toAddress && (ens?.clearRecords || ens?.setAddress || ens?.transferControl)) {
-        const transferENSResult = await transferENS({
-          clearRecords: ens.clearRecords,
-          name: ensName,
-          records: {
-            ...(ensProfile?.data?.contenthash ? { contenthash: ensProfile?.data?.contenthash } : {}),
-            ...(ensProfile?.data?.records || {}),
-            ...(ensProfile?.data?.coinAddresses || {}),
-          },
-          setAddress: ens.setAddress,
-          toAddress,
-          transferControl: ens.transferControl,
-          wallet,
-        });
-
-        if (!transferENSResult) {
-          logger.error(new RainbowError(`[SendSheet]: transferENS failed`), {
-            transferENSResult,
-          });
-          return;
-        }
-
-        if (typeof transferENSResult.nonce === 'number') {
-          nextNonce = transferENSResult.nonce + 1;
-        }
-      }
-
-      const gasLimitToUse = updatedGasLimit && !lessThan(updatedGasLimit, gasLimit) ? updatedGasLimit : gasLimit;
-
-      const gasParams = parseGasParamsForTransaction(selectedGasFee);
-      const txDetails: Partial<NewTransaction> = {
-        amount: amountDetails.assetAmount,
-        asset: selected as ParsedAddressAsset,
-        from: accountAddress,
-        gasLimit: gasLimitToUse,
-        network: currentChainIdNetwork,
-        chainId: currentChainId,
-        nonce: nextNonce ?? (await getNextNonce({ address: accountAddress, chainId: currentChainId })),
-        to: toAddress,
-        ...gasParams,
-      };
-
-      try {
-        const signableTransaction = await executeFn(createSignableTransaction, {
-          operation: TimeToSignOperation.CreateSignableTransaction,
-          screen: isENS ? Screens.SEND_ENS : Screens.SEND,
-        })(txDetails as NewTransactionNonNullable);
-        if (!signableTransaction.to) {
-          logger.error(new RainbowError(`[SendSheet]: txDetails is missing the "to" field`), {
-            txDetails,
-            signableTransaction,
-          });
-          Alert.alert(i18n.t(i18n.l.wallet.transaction.alert.invalid_transaction));
-          submitSuccess = false;
-        } else {
-          const sendTransactionResult = await executeFn(sendTransaction, {
-            screen: isENS ? Screens.SEND_ENS : Screens.SEND,
-            operation: TimeToSignOperation.BroadcastTransaction,
-          })({
-            existingWallet: wallet,
-            provider: currentProvider,
-            transaction: {
-              ...signableTransaction,
-              to: signableTransaction.to,
-              data: signableTransaction.data,
-              from: signableTransaction.from,
-              gasLimit: signableTransaction.gasLimit,
-              chainId: signableTransaction.chainId as ChainId,
-              value: signableTransaction.value,
-              nonce: signableTransaction.nonce,
-            },
-          });
-
-          if (!sendTransactionResult || !sendTransactionResult.result) {
-            logger.error(new RainbowError(`[SendSheet]: No result from sendTransaction`), {
-              sendTransactionResult,
-              signableTransaction,
-            });
-            return;
-          }
-
-          if (sendTransactionResult?.error) {
-            logger.error(new RainbowError(`[SendSheet]: Error from sendTransaction`), {
-              sendTransactionResult,
-              signableTransaction,
-            });
-            return;
-          }
-
-          const { hash, nonce } = sendTransactionResult.result;
-          const { data, value } = signableTransaction;
-          if (!isEmpty(hash)) {
-            submitSuccess = true;
-
-            if (hash) {
-              txDetails.hash = hash;
-            }
-
-            if (data) {
-              txDetails.data = data;
-            }
-
-            if (value) {
-              txDetails.value = value;
-            }
-
-            txDetails.nonce = nonce;
-            txDetails.network = currentChainIdNetwork;
-            txDetails.chainId = currentChainId;
-            txDetails.txTo = signableTransaction.to;
-            txDetails.type = 'send';
-            txDetails.status = TransactionStatus.pending;
-            addNewTransaction({
-              address: accountAddress,
-              chainId: currentChainId,
-              transaction: txDetails as NewTransaction,
-            });
-
-            // Invalidate the interactions count query for this recipient. if not done,
-            // the cache time is 15 minutes so the number of interactions will not be updated
-            if (accountAddress && toAddress && nativeCurrency) {
-              queryClient.invalidateQueries(
-                interactionsCountQueryKey({
-                  fromAddress: accountAddress.toLowerCase() as Address,
-                  toAddress: toAddress.toLowerCase() as Address,
-                  currency: nativeCurrency,
-                })
-              );
-            }
-          }
-        }
-      } catch (error) {
-        submitSuccess = false;
-        logger.error(new RainbowError(`[SendSheet]: onSubmit error`), {
-          txDetails,
-          error,
-        });
-
-        // if hardware wallet, we need to tell hardware flow there was error
-        // have to check inverse or we trigger unwanted BT permissions requests
-        if (!(wallet instanceof Wallet)) {
-          setHardwareTXError(true);
-        }
-      }
-      return submitSuccess;
-    },
-    [
-      accountAddress,
-      amountDetails.assetAmount,
-      amountDetails.isSufficientBalance,
-      currentChainId,
-      currentProvider,
-      ensName,
-      ensProfile?.data?.coinAddresses,
-      ensProfile?.data?.contenthash,
-      ensProfile?.data?.records,
-      gasLimit,
-      isENS,
-      isSufficientGas,
-      isUniqueAsset,
-      isValidAddress,
-      isValidGas,
-      selected,
-      selectedGasFee,
-      toAddress,
-      transferENS,
-      updateTxFee,
-      updateTxFeeForOptimism,
-      queryClient,
-      nativeCurrency,
-    ]
-  );
-
-  const rainbowToastsEnabled = useRainbowToastEnabled();
-
-  const submitTransaction = useCallback(
-    async (args?: OnSubmitProps) => {
-      if (Number(amountDetails.assetAmount) <= 0) {
-        logger.error(new RainbowError(`[SendSheet]: preventing tx submit because amountDetails.assetAmount is <= 0`), {
-          amountDetails,
-        });
-        return false;
-      }
-      const submitSuccessful = await onSubmit(args);
-      analytics.track(analytics.event.sentTransaction, {
-        assetName: selected?.name || '',
-        network: selected?.network || '',
-        isRecepientENS: recipient.slice(-4).toLowerCase() === '.eth',
-        isHardwareWallet,
-      });
-
-      const goBackAndNavigate = () => {
-        goBack();
-        navigate(Routes.WALLET_SCREEN);
-
-        // with toasts: just show the toast
-        // without toasts: navigate user to the activity list
-        if (!rainbowToastsEnabled) {
-          InteractionManager.runAfterInteractions(() => {
-            navigate(Routes.PROFILE_SCREEN);
-          });
-        }
-      };
-
-      if (submitSuccessful) {
-        // if the user sent an NFT, we need to revalidate the NFT data
-        if (isUniqueAsset) {
-          const collectionId = `${selected.network}_${selected.contractAddress}`;
-          useNftsStore.getState(accountAddress).fetchNftCollection(collectionId, true);
-          useNftsStore.getState(accountAddress).fetch({ limit: PAGE_SIZE }, { staleTime: time.seconds(5) });
-        }
-        executeFn(goBackAndNavigate, {
-          screen: isENS ? Screens.SEND_ENS : Screens.SEND,
-          operation: TimeToSignOperation.SheetDismissal,
-          isEndOfFlow: true,
-        })();
-      }
-    },
-    [
-      accountAddress,
-      amountDetails,
-      goBack,
-      isENS,
-      isHardwareWallet,
-      isUniqueAsset,
-      navigate,
-      onSubmit,
-      rainbowToastsEnabled,
-      recipient,
-      selected,
-    ]
-  );
+  const { submitTransaction } = useSendSubmit({
+    accountAddress,
+    amountDetails,
+    currentChainId,
+    currentProvider,
+    ens: { ensName, ensProfile, isENS, transferENS },
+    gas: { gasLimit, isSufficientGas, isValidGas, selectedGasFee, updateTxFee, updateTxFeeForOptimism },
+    isHardwareWallet,
+    isUniqueAsset,
+    nativeCurrency,
+    recipient: { isValidAddress, recipient, toAddress },
+    selected,
+  });
 
   const { buttonDisabled, buttonLabel } = useMemo(() => {
-    const isZeroAssetAmount = Number(amountDetails.assetAmount) <= 0;
-    let disabled = true;
-    let label = i18n.t(i18n.l.button.confirm_exchange.enter_amount);
+    const isGasFeeReady =
+      !isEmpty(gasFeeParamsBySpeed) &&
+      Boolean(selectedGasFee) &&
+      !isEmpty(selectedGasFee?.gasFee) &&
+      Boolean(toAddress) &&
+      (!useBackendNetworksStore.getState().getNeedsL1SecurityFeeChains().includes(currentChainId) || l1GasFeeOptimism !== null);
 
-    if (isENS && !ensProfile.isSuccess) {
-      label = i18n.t(i18n.l.button.confirm_exchange.loading);
-      disabled = true;
-    } else if (
-      isEmpty(gasFeeParamsBySpeed) ||
-      !selectedGasFee ||
-      isEmpty(selectedGasFee?.gasFee) ||
-      !toAddress ||
-      (useBackendNetworksStore.getState().getNeedsL1SecurityFeeChains().includes(currentChainId) && l1GasFeeOptimism === null)
-    ) {
-      label = i18n.t(i18n.l.button.confirm_exchange.loading);
-      disabled = true;
-    } else if (!isZeroAssetAmount && !isSufficientGas) {
-      disabled = true;
-      label = i18n.t(i18n.l.button.confirm_exchange.insufficient_token, {
-        tokenName: useBackendNetworksStore.getState().getChainsNativeAsset()[currentChainId || ChainId.mainnet]?.symbol,
-      });
-    } else if (!isValidGas) {
-      disabled = true;
-      label = i18n.t(i18n.l.button.confirm_exchange.invalid_fee);
-    } else if (!isZeroAssetAmount && !amountDetails.isSufficientBalance) {
-      disabled = true;
-      label = i18n.t(i18n.l.button.confirm_exchange.insufficient_funds);
-    } else if (!isZeroAssetAmount) {
-      disabled = false;
-      label = `􀕹 ${i18n.t(i18n.l.button.confirm_exchange.review)}`;
-    }
-
-    return { buttonDisabled: disabled, buttonLabel: label };
+    return getSendSubmitButtonState({
+      assetAmount: amountDetails.assetAmount,
+      isENS,
+      isENSProfileLoaded: ensProfile.isSuccess,
+      isGasFeeReady,
+      isSufficientBalance: amountDetails.isSufficientBalance,
+      isSufficientGas,
+      isValidGas,
+      nativeAssetSymbol: useBackendNetworksStore.getState().getChainsNativeAsset()[currentChainId || ChainId.mainnet]?.symbol,
+    });
   }, [
     amountDetails.assetAmount,
     amountDetails.isSufficientBalance,
